+++ spec/dummy/config/routes.rb
@@ -7,6 +7,7 @@ Rails.application.routes.draw do
       get :filters
       get :search
       get :checkboxes_dropdown
+      match :modals, via: [:get, :post]
     end
   end
 end