+++ features/support/env.rb
@@ -1,3 +1,6 @@
 ENV["RAILS_ROOT"] ||= File.expand_path("../../spec/dummy", __dir__)
 require "cucumber/rails"
 require "agilibox/cucumber_config"
+Agilibox::CucumberConfig.require_all_helpers!
+Agilibox::CucumberConfig.require_poltergeist!
+# Agilibox::CucumberConfig.require_chrome_headless!